We can now create plugins and display output to a view which is rendered in a semi-complete theme. The routing system is also making leaps and bounds.
I'm very proud of the progress so far. We still have a little way to go, but right now I've ported a theme into AllianceCMS, and we can create a plugin that will allow us to define routes which will call our controller, which will pass data to our plugin views, which will then be displayed in the ported theme. We finally have something to look at and play with after 2+ weeks of working on back end code :-)