Now dynamically loads 'dbConnections.php'. The system reads the server name and checks if there is a folder with the same name in '/zones'. This is great for different db connections/databases for dev/staging/production phases.
If it is the main domain (mysite.com) it will look for the folder mysite.com. If the folder mysite.com has a 'dbConnections.php' then it will load it.  If that folder doesn't exist it will load the '/zones/default/dbConnections.php'. If that doesn't exist the site will start the installation.

If the site is on a subdomain (subdomain.mysite.com), the system will only look for '/zones/subdomain.mysite.com/dbConnections.php'. If that does not exist then the site will start the installation.